Update Auth.html
I was getting errors in the console on Auth.html until I prefixed Log and UserManager with '.Oidc.' I am not sure if this is due to the oidc-client version I am using (v1.3.0) or if the issue is in the code on my end. I noticed in these docs they are prefixing with '.Oidc' so I tried it and the errors went away.